addon/rootfs/usr/src/app/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
httpx==0.27.2
websockets==13.1
pydantic==2.9.2
python-dotenv==1.0.1
